authorization

+++ Controller/Customer/deleteCustomer.js
@@ -3,12 +3,15 @@ const Customer = require('../../Model/Customer');
 const deleteCustomer = async (req, res) => {
   try {
     const customerId = req.body.Customerid;
+    const adminId = req.user.adminId;
+
+    const customer = await Customer.findOne({ _id: customerId, AdminID: adminId });
+
+    if (!customer) {
+      return res.status(404).json({ message: 'Customer not found or not authorized' });
+    }
 
     const deletedCustomer = await Customer.findByIdAndDelete(customerId);
-
-    if (!deletedCustomer) {
-      return res.status(404).json({ message: 'Customer not found' });
-    }
 
     res.status(200).json({
       message: 'Customer deleted successfully',
@@ -20,5 +23,4 @@ const deleteCustomer = async (req, res) => {
   }
 };
 
-
-module.exports={deleteCustomer}
+module.exports = { deleteCustomer };
