+++ Controller/Invoice/deleteInvoice.js
@@ -3,12 +3,19 @@ const Invoice = require('../../Model/Invoices');
 const deleteInvoice = async (req, res) => {
   try {
     const { invoiceId } = req.body;
+    const adminId = req.user.adminId;
+
+    if (!invoiceId) {
+      return res.status(400).json({ message: 'Invoice ID is required' });
+    }
+
+    const invoice = await Invoice.findOne({ _id: invoiceId, AdminID: adminId });
+
+    if (!invoice) {
+      return res.status(404).json({ message: 'Invoice not found or not authorized' });
+    }
 
     const deletedInvoice = await Invoice.findByIdAndDelete(invoiceId);
-
-    if (!deletedInvoice) {
-      return res.status(404).json({ message: 'Invoice not found' });
-    }
 
     res.status(200).json({
       message: 'Invoice deleted successfully',
@@ -19,4 +26,5 @@ const deleteInvoice = async (req, res) => {
     res.status(500).json({ message: 'Internal server error' });
   }
 };
-module.exports={deleteInvoice}
+
+module.exports = { deleteInvoice };
